ROX-31581: Rewrite APITokensService in TypeScript
